Create routes folder, turn off prefer default exports, remove default exports
