+++ .eslintrc.json
@@ -21,6 +21,7 @@
   },
   "plugins": ["react", "@typescript-eslint", "prettier", "cypress", "unused-imports"],
   "rules": {
+    "import/prefer-default-export": "off",
     "unused-imports/no-unused-imports": "error",
     "no-use-before-define": "off",
     "@typescript-eslint/no-use-before-define": ["error"],
