+++ src/routes/DefaultRoute.tsx
@@ -1,0 +1,26 @@
+import React, { useContext } from 'react';
+import { Redirect, Route } from 'react-router-dom';
+import { AuthContext } from '../context/AuthProvider';
+
+const DefaultRoute = () => {
+  const user = useContext(AuthContext);
+
+  return (
+    <Route
+      path="/"
+      render={() => {
+        if (user === undefined) {
+          return <p>Loading...</p>;
+        }
+
+        if (user) {
+          return <Redirect to="/feed" />;
+        }
+
+        return <Redirect to="/" />;
+      }}
+    />
+  );
+};
+
+export { DefaultRoute };
