+++ src/routes/protectedDefaultRoutes.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+import { DefaultRoute } from './DefaultRoute';
+import { ProtectedRoute } from './ProtectedRoute';
+
+const protectedDefaultRoutes = [<ProtectedRoute path="/feed" />, <DefaultRoute />];
+
+export { protectedDefaultRoutes };
